+++ src/Cht/Message/Reply/UsingLlm/Engine.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Ssc\Btlr\Cht\Augment\UsingLlm\Model;
+namespace Ssc\Btlr\Cht\Message\Reply\UsingLlm;
 
 interface Engine
 {