+++ src/Cht/Message/Reply/UsingLlm/Engine/Cli.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Ssc\Btlr\Cht\Augment\UsingLlm\Model\Engine;
+namespace Ssc\Btlr\Cht\Message\Reply\UsingLlm\Engine;
 
 use Ssc\Btlr\App\Stdio;
 use Ssc\Btlr\App\Stdio\Write\WithStyle;
-use Ssc\Btlr\Cht\Augment\UsingLlm\Model\Engine;
+use Ssc\Btlr\Cht\Message\Reply\UsingLlm\Engine;
 
 class Cli implements Engine
 {