+++ src/Cht/Message/Reply/UsingLlm/Engine/Dummy.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Ssc\Btlr\Cht\Augment\UsingLlm\Model\Engine;
+namespace Ssc\Btlr\Cht\Message\Reply\UsingLlm\Engine;
 
-use Ssc\Btlr\Cht\Augment\UsingLlm\Model\Engine;
+use Ssc\Btlr\Cht\Message\Reply\UsingLlm\Engine;
 
 class Dummy implements Engine
 {